+++ src/test/kotlin/com/atlassian/performance/tools/ssh/api/SshContainer.kt
@@ -4,8 +4,8 @@ package com.atlassian.performance.tools.ssh.api
 import com.atlassian.performance.tools.ssh.api.auth.PublicKeyAuthentication
 import com.atlassian.performance.tools.sshubuntu.api.SshUbuntuContainer
 
-class SshContainer {
-    fun run(action: (ssh: SshConnection) -> Unit) {
+internal class SshContainer {
+    internal fun useConnection(action: (sshConnection: SshConnection) -> Unit) {
         SshUbuntuContainer().start().use { sshUbuntu ->
             return@use Ssh(
                 sshUbuntu.ssh.toSshHost()
@@ -13,6 +13,13 @@ class SshContainer {
                 .use { action(it) }
         }
     }
+
+    internal fun useSsh(action: (ssh: Ssh) -> Unit) {
+        SshUbuntuContainer().start().use { sshUbuntu ->
+            action(Ssh(sshUbuntu.ssh.toSshHost()))
+        }
+    }
+
 
     private fun com.atlassian.performance.tools.sshubuntu.api.SshHost.toSshHost(): SshHost {
         return SshHost(