+++ CHANGELOG.md
@@ -18,6 +18,10 @@ The API consists of all public Kotlin types from `com.atlassian.performance.tool
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/ssh/branches/compare/master%0Drelease-2.1.0
 
+### Added
+- Support local port forwarding.
+- Support remote port forwarding.
+
 ## [2.1.0] - 2018-10-26
 [2.1.0]: https://bitbucket.org/atlassian/ssh/branches/compare/release-2.1.0%0Drelease-2.0.0